Broke down members settings
