+++ app/templates/settings/members-payments.hbs
@@ -1,9 +1,7 @@
 <section class="gh-canvas">
     <GhCanvasHeader class="gh-canvas-header">
         <h2 class="gh-canvas-title" data-test-screen-title>
-            <LinkTo @route="settings">Settings</LinkTo>
-            <span>{{svg-jar "arrow-right"}}</span>
-            Members
+            Payments
         </h2>
         <section class="view-actions">
             <GhTaskButton @buttonText="Save settings"
@@ -35,13 +33,10 @@
 
         <div class="gh-setting-liquid-section">
             {{#liquid-if this.feature.labs.members}}
-            <GhMembersLabSetting
+            <GhMembersPaymentsSetting
                 @settings={{this.settings}}
-                @fromAddress={{this.fromAddress}}
-                @supportAddress={{this.supportAddress}}
                 @setDefaultContentVisibility={{action "setDefaultContentVisibility"}}
                 @setStripeConnectIntegrationTokenSetting={{action "setStripeConnectIntegrationTokenSetting"}}
-                @setEmailAddress={{action "setEmailAddress"}}
             />
             {{/liquid-if}}
         </div>
